fixed subcommands not showing up in help embed (totally all me)
