+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "iris",
-  "version": "2-DEV-11b4b1e5",
+  "version": "2-DEV-ac16968f",
   "description": "IRIS (Intelligent Response Interface System) is a Discord bot created by and for the Chaos Crew Community.",
   "main": "index.js",
   "scripts": {
